switching internally to use double instead of float in TriangleIntersector.intersectsRay()
--HG--
branch : toxiclibs

+++ src.core/toxi/geom/TriangleIntersector.java
@@ -60,16 +60,16 @@ public class TriangleIntersector implements Intersector3D {
         float dotprod = n.dot(ray.dir);
         if (dotprod < 0) {
             Vec3D rt = ray.sub(triangle.a);
-            float t = -(n.x * rt.x + n.y * rt.y + n.z * rt.z)
+            double t = -(double) (n.x * rt.x + n.y * rt.y + n.z * rt.z)
                     / (n.x * ray.dir.x + n.y * ray.dir.y + n.z * ray.dir.z);
             if (t >= MathUtils.EPS) {
-                Vec3D pos = ray.getPointAtDistance(t);
+                Vec3D pos = ray.getPointAtDistance((float) t);
                 // check if pos is inside triangle
                 if (triangle.containsPoint(pos)) {
                     isectData.isIntersection = true;
                     isectData.pos = pos;
                     isectData.normal = n;
-                    isectData.dist = t;
+                    isectData.dist = (float) t;
                     isectData.dir = isectData.pos.sub(ray).normalize();
                 }
             }
